meson rewrite
