+++ .gitignore
@@ -1,2 +1,1 @@
-src/QxtNetwork/*
-src/QxtWeb/*
+build*/*
