+++ src/QxtWeb/HttpSessionManager
@@ -1,0 +1,1 @@
+#include "web/qxthttpsessionmanager.h"
