+++ src/QxtWeb/QxtHttpServerConnector
@@ -1,0 +1,1 @@
+#include "qxtabstracthttpconnector.h"
