+++ src/QxtWeb/QxtHttpSessionManager
@@ -1,0 +1,1 @@
+#include "web/qxthttpsessionmanager.h"
