+++ src/QxtWeb/QxtWebContent
@@ -1,0 +1,1 @@
+#include "web/qxtwebcontent.h"
