+++ src/QxtWeb/QxtWebPageEvent
@@ -1,0 +1,1 @@
+#include "web/qxtwebevent.h"
