+++ src/QxtWeb/QxtWebSlotService
@@ -1,0 +1,1 @@
+#include "web/qxtwebslotservice.h"
