--- src/README.txt
@@ -1,1 +1,0 @@
-QxtWeb from libqxt.org project, standalone and cmakeified. For embeddable webserver
