+++ src/core/qxtboundfunctionbase.h
