+++ src/core/qxtglobal.h
