+++ src/core/qxtmetaobject.cpp
