+++ src/core/qxtnull.cpp
