+++ src/core/qxtnull.h
