+++ src/core/qxtnullable.h
