+++ src/web/qxtabstracthttpconnector.cpp
