+++ src/web/qxtabstracthttpconnector.h
