+++ src/web/qxthtmltemplate.cpp
