+++ src/web/qxthttpserverconnector.cpp
