+++ src/web/qxthttpsessionmanager.cpp
