+++ src/web/qxthttpsessionmanager.h
